Render AuthButton as child
